requirements.txt
python-dotenv==0.20.*
python-telegram-bot==13.*
requests==2.28.*